Chore[adreno_tools]: move DriverConfig into new package
